Refactor Multi-Configuration tests with POM structure
